Create RAG_Company_IT_Policy
The key to solving this problem is to retrieve the original paragraphs in the document that are related to the question, and use the Nearest Neighbor Search method. Implemented with Python scripts - cut the document into paragraphs and convert it into a numpy two-dimensional array using the paraphrase-MiniLM-L6-v2 model and store it in the faiss.IndexFlatL2 searcher. Convert the user-entered question into a numpy two-dimensional array, retrieve the index of the paragraph vector closest to the question vector, and then find the corresponding paragraph. Write the role setting, instruction description, paragraph and question into ragprompt.txt. Use the content in the text as a prompt to start the Qwen reasoning program to get the answer.